FIX short

+++ debezium-core/src/main/java/io/debezium/jdbc/JdbcValueConverters.java
@@ -786,7 +786,7 @@ public class JdbcValueConverters implements ValueConverterProvider {
         }
         if (data == null) {
             if (column.isOptional()) return null;
-            return (short)0;
+            return Short.valueOf((short) 0);
         }
         if (data instanceof Short) return data;
         if (data instanceof Number) {
